We need to get the healpix package from ubuntu utopic
Because travis is on ubuntu precise, we need to download the source and
build the package from scratch.

There does seem to be an old ppa with the right packages
(https://launchpad.net/~zonca/+archive/ubuntu/healpix), but I couldn't
dlopen the shared library after installing from here. It was weird.

Debugging travis is on my top ten list of things I hate. This one took
28 tries to get right :(