the autotools build system for healpix builds a shared library
If you just do the dumb `./configure && make` you only get
libhealpix_cxx.a where as now we're getting libhealpix_cxx.so as well.
The latter seems to be preferred by the linker so we need to dlopen the
shared library to resolve some of the symbols in libhealpixwrapper.so

I think this is the right explanation anyway. I'm not so good at this
build system stuff.