make sure cfitsio is installed, this is also a dependency!
Apparently we cannot rely on FITSIO.jl to install cfitsio for us. I
think that is sometimes getting built in a local directory and not being
added to the appropriate paths. There might be a more general solution
to this problem but for now we'll just ask apt to install the package
for us.

TODO: a solution for macs, and a build from source