widen the signature on most pixel functions
Now we'll just try to convert each argument to Clong and call the Clong
version in libchealpix. We no longer also try to wrap the Clonglong
version of each function.

This means, for example, the `nside` and `pixel` arguments no longer
have to have the same type and we can accept that `query_disc` returns a
`Vector{Cint}` instead of a `Vector{Int}`.